refactor : 동시성 테스트 분리 및 필요한 데이터 쿼리 작성
