+++ build.gradle
@@ -33,6 +33,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-jdbc:3.1.1'
     implementation 'org.springframework.vault:spring-vault-core:3.0.2'
+    implementation group: 'com.github.javafaker', name: 'javafaker', version: '0.16'
 
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
