+++ src/test/java/com/heeverse/member/domain/MemberTestHelper.java
@@ -1,7 +1,9 @@
 package com.heeverse.member.domain;
 
+import com.github.javafaker.Faker;
 import com.heeverse.member.domain.entity.Member;
-import com.heeverse.member.dto.MemberRequestDto;
+
+import java.util.Locale;
 
 /**
  * @author jeongheekim
@@ -9,15 +11,15 @@ import com.heeverse.member.dto.MemberRequestDto;
  */
 public class MemberTestHelper {
     public static final String MEMBER_ID = "guestzzang";
-    public static final String NAME = "armyBTS";
+
     public static final String PASSWORD = "Helloworld!123";
-    public static final String EMAIL = "guestzzang@gmail.com";
 
     private MemberTestHelper() {
     }
 
     public static Member getMockMember() {
-        return new Member(MEMBER_ID, NAME, PASSWORD, EMAIL);
+        Faker faker = new Faker(Locale.US);
+        return new Member(MEMBER_ID + faker.number().randomDigit(), PASSWORD, faker.name().name(), faker.internet().emailAddress());
     }
 
 }
